# src/main/java/com/effective_mobile/card_management/dto/AuthenticationResponse.java
package com.effective_mobile.card_management.dto;
import lombok.Getter;

import java.io.Serializable;

@Getter
public class AuthenticationResponse implements Serializable {

    private final String jwt;

    public AuthenticationResponse(String jwt) {
        this.jwt = jwt;
    }
}